Changed password to be a SHA-256 hash.
Changed balance fields to double

+++ src/main/java/com/stocktrader/controllers/DepositViewController.java
@@ -13,7 +13,7 @@ public class DepositViewController {
 
     @FXML
     private void deposit() {
-        int amount = Integer.parseInt(depositAmountTextField.getText());
+        double amount = Double.parseDouble(depositAmountTextField.getText());
         new Deposit(username, amount);
         welcomePageController.displayPortfolio();
         welcomePageController.closeDepositWindow();
